+++ shared/chat/conversation/input-area/normal/platform-input.desktop.js
