+++ shared/chat/conversation/input-area/normal/platform-input.js.flow
